pyproject.toml
[tool.poetry]
name = "dbvirus"
version = "0.1.0"
description = ""
authors = ["Felipe Rodrigues <felipe@felipevr.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.7"
biopython = "^1.74"
xmltodict = "^0.12"
mongoengine = "^0.18.2"
fire = "^0.2.1"

[tool.poetry.dev-dependencies]
pylint = "^2.3"
black = {version = "^18.3-alpha.0",allows-prereleases = true}
pytest = "^5.0"
pytest-mock = "^1.10"
jsonschema = "^3.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
